Fix with_db integration tests
