+++ tests/integration_tests/with_db/udfs/test_question_answering_script.py
@@ -1,3 +1,4 @@
+from tests.integration_tests.with_db.udfs.python_rows_to_sql import python_rows_to_sql
 from tests.utils.parameters import model_params
 
 
@@ -12,6 +13,7 @@ def test_question_answering_script(
         input_data.append((
             '',
             bucketfs_conn_name,
+            None,
             str(model_params.sub_dir),
             model_params.base_model,
             question,
@@ -22,13 +24,14 @@ def test_question_answering_script(
     query = f"SELECT TE_QUESTION_ANSWERING_UDF(" \
             f"t.device_id, " \
             f"t.bucketfs_conn_name, " \
+            f"t.token_conn_name, " \
             f"t.sub_dir, " \
             f"t.model_name, " \
             f"t.question, " \
             f"t.context_text, " \
             f"t.top_k" \
-            f") FROM (VALUES {str(tuple(input_data))} " \
-            f"AS t(device_id, bucketfs_conn_name, sub_dir, " \
+            f") FROM (VALUES {python_rows_to_sql(input_data)} " \
+            f"AS t(device_id, bucketfs_conn_name, token_conn_name, sub_dir, " \
             f"model_name, question, context_text, top_k));"
 
     # execute sequence classification UDF
