+++ tests/integration_tests/with_db/udfs/test_sequence_classification_text_pair_script.py
@@ -1,3 +1,4 @@
+from tests.integration_tests.with_db.udfs.python_rows_to_sql import python_rows_to_sql
 from tests.utils.parameters import model_params
 
 
@@ -11,6 +12,7 @@ def test_sequence_classification_text_pair_script(
         input_data.append((
             '',
             bucketfs_conn_name,
+            None,
             str(model_params.sub_dir),
             model_params.base_model,
             model_params.text_data,
@@ -19,12 +21,13 @@ def test_sequence_classification_text_pair_script(
     query = f"SELECT TE_SEQUENCE_CLASSIFICATION_TEXT_PAIR_UDF(" \
             f"t.device_id, " \
             f"t.bucketfs_conn_name, " \
+            f"t.token_conn_name, " \
             f"t.sub_dir, " \
             f"t.model_name, " \
             f"t.first_text, " \
             f"t.second_text" \
-            f") FROM (VALUES {str(tuple(input_data))} " \
-            f"AS t(device_id, bucketfs_conn_name, sub_dir, " \
+            f") FROM (VALUES {python_rows_to_sql(input_data)} " \
+            f"AS t(device_id, bucketfs_conn_name, token_conn_name, sub_dir, " \
             f"model_name, first_text, second_text));"
 
     # execute sequence classification UDF
