+++ tests/integration_tests/with_db/udfs/test_text_generation_script.py
@@ -1,3 +1,4 @@
+from tests.integration_tests.with_db.udfs.python_rows_to_sql import python_rows_to_sql
 from tests.utils.parameters import model_params
 
 
@@ -13,6 +14,7 @@ def test_text_generation_script(
         input_data.append((
             '',
             bucketfs_conn_name,
+            None,
             str(model_params.sub_dir),
             model_params.base_model,
             text_data,
@@ -23,13 +25,14 @@ def test_text_generation_script(
     query = f"SELECT TE_TEXT_GENERATION_UDF(" \
             f"t.device_id, " \
             f"t.bucketfs_conn_name, " \
+            f"t.token_conn_name, " \
             f"t.sub_dir, " \
             f"t.model_name, " \
             f"t.text_data, " \
             f"t.max_length," \
             f"t.return_full_text" \
-            f") FROM (VALUES {str(tuple(input_data))} " \
-            f"AS t(device_id, bucketfs_conn_name, sub_dir, " \
+            f") FROM (VALUES {python_rows_to_sql(input_data)} " \
+            f"AS t(device_id, bucketfs_conn_name, token_conn_name, sub_dir, " \
             f"model_name, text_data, max_length, return_full_text));"
 
     # execute sequence classification UDF
