+++ tests/integration_tests/with_db/udfs/test_token_classification_script.py
@@ -1,3 +1,4 @@
+from tests.integration_tests.with_db.udfs.python_rows_to_sql import python_rows_to_sql
 from tests.utils.parameters import model_params
 
 
@@ -11,6 +12,7 @@ def test_token_classification_script(
         input_data.append((
             '',
             bucketfs_conn_name,
+            None,
             str(model_params.sub_dir),
             model_params.base_model,
             model_params.text_data,
@@ -20,12 +22,13 @@ def test_token_classification_script(
     query = f"SELECT TE_TOKEN_CLASSIFICATION_UDF(" \
             f"t.device_id, " \
             f"t.bucketfs_conn_name, " \
+            f"t.token_conn_name, " \
             f"t.sub_dir, " \
             f"t.model_name, " \
             f"t.text_data, " \
             f"t.aggregation_strategy" \
-            f") FROM (VALUES {str(tuple(input_data))} " \
-            f"AS t(device_id, bucketfs_conn_name, " \
+            f") FROM (VALUES {python_rows_to_sql(input_data)} " \
+            f"AS t(device_id, bucketfs_conn_name, token_conn_name, " \
             f"sub_dir, model_name, text_data, aggregation_strategy));"
 
     # execute sequence classification UDF
